Remove the timeline component as there is now only one timeline/page + delete art and politics pages
